souvenir/requirements.txt
Django==1.10.4
django-guardian==1.4.6
django-materialize-css==0.0.1
django-userena==2.0.1
easy-thumbnails==2.3
html2text==2014.12.29
Pillow==3.4.2
six==1.10.0
